requirements.txt
Django>=4.2,<5.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
